// frontend-ui/src/app/services/models/shareholder-details-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ShareholderDto } from '../models/shareholder-dto';

/**
 * Shareholder details DTO
 */
export interface ShareholderDetailsDto {

  /**
   * Active since date
   */
  activeSince?: string;

  /**
   * Current balance
   */
  currentBalance?: number;

  /**
   * Investment amount
   */
  investment?: number;
  shareholder?: ShareholderDto;

  /**
   * Total earnings
   */
  totalEarnings?: number;

  /**
   * Total shares
   */
  totalShares?: number;

  /**
   * Total value
   */
  totalValue?: number;
}